+++ test/state_flow/test_helpers.clj
@@ -1,5 +1,6 @@
 (ns state-flow.test-helpers
-  (:require [state-flow.core]
+  (:require [state-flow.cljtest :as cljtest]
+            [state-flow.core]
             [state-flow.state :as state]))
 
 (defmacro this-line-number
@@ -32,28 +33,3 @@
   `(with-redefs [clojure.test/do-report identity]
      (binding [*out* (clojure.java.io/writer (java.io.File/createTempFile "test" "log"))]
        ~@body)))
-
-(defmacro run-flow
-  "Wrapper for `state-flow.core/run!`, but captures clojure.test's report data
-  instead of printing it to *out*. Returns a map of:
-
-    :report-data - the data that _would_ be reported via clojure.test
-    :flow-ret    - the return value of the flow
-    :flow-state  - the end-state of the flow "
-  [flow & [state]]
-  `(let [report-data# (atom nil)
-         res#         (with-redefs [clojure.test/do-report (fn [data#] (reset! report-data# data#))]
-                        (binding [*out* (clojure.java.io/writer (java.io.File/createTempFile "test" "log"))]
-                          (state-flow.core/run!
-                           ~flow
-                           ~(or state {}))))]
-     {:report-data (->> (deref report-data#)
-                        ;; NOTE: :matcher-combinators.result/value is a Mismatch object, which is
-                        ;; a defrecord, so equality on a map won't pass, hence pouring it into a map
-                        ;; to facilitate equality checks.
-                        (clojure.walk/postwalk (fn [node#]
-                                                 (if (instance? matcher_combinators.model.Mismatch node#)
-                                                   (into {} node#)
-                                                   node#))))
-      :flow-ret    (first res#)
-      :flow-state  (second res#)}))